Initial change-set for HBase0.98 -- Sidebranch Only!!!
NOT backward compatible with earlier HBase
For now, this is going into hbase98 branch for testing purposes.

Author: Oliver Bucaojit <oliver.bucaojit@hp.com>
Author: Narendra Goyal <narendra.goyal@hp.com>
Author: Joanie Cooper <joanie.cooper@hp.com>
Author: Shang-Sheng Tung <shang-sheng.tung@hp.com>
Author: Mike Hanlon <mike.hanlon@hp.com>
Author: Khaled Bouaziz <khaled.bouaziz@hp.com>
Author: Prashanth Vasudev <vasudev.prashanth@hp.com>

+++ sql/executor/HBaseClient.java
@@ -18,12 +18,12 @@
 
 package org.trafodion.sql.HBaseAccess;
 
+import com.google.protobuf.ServiceException;
 import java.io.IOException;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
 import java.util.ArrayList;
-
 
 import org.apache.log4j.PropertyConfigurator;
 import org.apache.log4j.Logger;
@@ -32,6 +32,8 @@ import org.apache.hadoop.hbase.HBaseConfiguration;
 import org.apache.hadoop.hbase.HTableDescriptor;
 import org.apache.hadoop.hbase.HColumnDescriptor;
 import org.apache.hadoop.hbase.KeyValue;
+import org.apache.hadoop.hbase.NamespaceDescriptor;
+import org.apache.hadoop.hbase.TableName;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.client.HBaseAdmin;
 import org.apache.hadoop.hbase.util.Bytes;
@@ -49,11 +51,11 @@ import org.apache.hadoop.hbase.protobuf.generated.HBaseProtos.SnapshotDescriptio
 import org.apache.hadoop.hbase.protobuf.generated.HBaseProtos.SnapshotDescription.Type;
 //import org.apache.hadoop.hbase.protobuf.generated.HBaseProtos;
 
-//import org.apache.hadoop.hbase.io.compress.Compression.Algorithm;
-import org.apache.hadoop.hbase.io.hfile.Compression.Algorithm;
+import org.apache.hadoop.hbase.io.compress.Compression.Algorithm;
+//import org.apache.hadoop.hbase.io.hfile.Compression.Algorithm;
 import org.apache.hadoop.hbase.io.encoding.DataBlockEncoding;
-//import org.apache.hadoop.hbase.regionserver.BloomType; in v0.97
-import org.apache.hadoop.hbase.regionserver.StoreFile.BloomType ;
+import org.apache.hadoop.hbase.regionserver.BloomType; 
+//import org.apache.hadoop.hbase.regionserver.StoreFile.BloomType ;
 import org.apache.hadoop.hbase.regionserver.KeyPrefixRegionSplitPolicy;
 //import org.apache.hadoop.hbase.client.Durability;
 import org.trafodion.sql.HBaseAccess.StringArrayList;
@@ -118,7 +120,8 @@ public class HBaseClient {
     }
 
     public boolean init(String zkServers, String zkPort) 
-              throws MasterNotRunningException, ZooKeeperConnectionException {
+	throws MasterNotRunningException, ZooKeeperConnectionException, ServiceException, IOException
+    {
          setupLog4j();
          logger.debug("HBaseClient.init(" + zkServers + ", " + zkPort
                          + ") called.");
@@ -391,8 +394,6 @@ public class HBaseClient {
             logger.debug("HBaseClient.dropAll(" + pattern + ") called.");
             HBaseAdmin admin = new HBaseAdmin(config);
 
-	    //	    System.out.println(pattern);
-
 	    HTableDescriptor[] htdl = admin.listTables(pattern);
 	    if (htdl == null) // no tables match the given pattern.
 		return true;
@@ -400,27 +401,16 @@ public class HBaseClient {
 	    for (HTableDescriptor htd : htdl) {
 		String tblName = htd.getNameAsString();
 
-		//		System.out.println(tblName);
-		admin.disableTable(tblName);
-		admin.deleteTable(tblName);
+                // do not drop DTM log files which have the format: TRAFODION._DTM_.*
+                int idx = tblName.indexOf("TRAFODION._DTM_");
+                if (idx == 0)
+                    continue;
+                
+                //                System.out.println(tblName);
+                admin.disableTable(tblName);
+                admin.deleteTable(tblName);
 	    }
  	    
-	    /*
-            HTableDescriptor[] htd = admin.disableTables(pattern);
-	    if (htd != null) {
-		System.out.println("here");
-		System.out.println(pattern);
-
-
-		return false;
-	    }
-
-            htd = admin.deleteTables(pattern);
-	    if (htd != null) {
-		return false;
-	    }
-	    */
-
             admin.close();
             return cleanup();
     }
@@ -542,11 +532,15 @@ public class HBaseClient {
                 i++;
             }
 
-            UserPermission userPerm = new UserPermission(user, tblName,
+	    //HB98
+	    TableName htblName = TableName.valueOf(new String(NamespaceDescriptor.DEFAULT_NAMESPACE_NAME)
+						   ,new String(tblName));
+            UserPermission userPerm = new UserPermission(user, htblName,
                                                          colFamily, assigned);
 
             AccessController accessController = new AccessController();
-            accessController.grant(userPerm);
+	    //HB98 The grant() method is very different in HB98 (commenting out for now)
+            //accessController.grant(userPerm);
         return true;
     }
 
@@ -569,11 +563,16 @@ public class HBaseClient {
                 i++;
             }
 
-            UserPermission userPerm = new UserPermission(user, tblName,
+	    //HB98
+	    TableName htblName = TableName.valueOf(new String(NamespaceDescriptor.DEFAULT_NAMESPACE_NAME)
+						   ,new String(tblName));
+            UserPermission userPerm = new UserPermission(user, htblName,
                                                          colFamily, assigned);
 
             AccessController accessController = new AccessController();
-            accessController.revoke(userPerm);
+	    
+	    //HB98 The revoke() method is very different in HB98 (commenting out for now)
+            //accessController.revoke(userPerm);
         return true;
     }
